requirements.txt
# ── Core AWS ────────────────────────────────────────────────────────────────
boto3>=1.34.0
awswrangler>=3.6.0

# ── ML ───────────────────────────────────────────────────────────────────────
xgboost>=2.0.0
scikit-learn>=1.4.0
pandas>=2.1.0
numpy>=1.26.0
shap>=0.44.0

# ── MLflow ───────────────────────────────────────────────────────────────────
mlflow>=2.11.0

# ── SageMaker ────────────────────────────────────────────────────────────────
sagemaker>=2.210.0

# ── RAG / Vector search ──────────────────────────────────────────────────────
opensearch-py>=2.4.0
requests-aws4auth>=1.2.3
pypdf>=4.1.0
langchain-text-splitters>=0.0.1

# ── Config / Utils ───────────────────────────────────────────────────────────
pyyaml>=6.0.1
python-dotenv>=1.0.0

# ── API serving ──────────────────────────────────────────────────────────────
fastapi>=0.110.0
uvicorn[standard]>=0.29.0

# ── Streamlit UI ─────────────────────────────────────────────────────────────
streamlit>=1.32.0
plotly>=5.20.0
pydeck>=0.8.0
